Move TopBar to root-level of the App for WindowControls and dragging
