+++ .eslintrc.js
@@ -3,7 +3,8 @@ module.exports = {
   rules: {
     // A temporary hack related to IDE not resolving correct package.json
     'import/no-extraneous-dependencies': 'off',
-    'jsx-a11y/anchor-is-valid': 0,
+    'jsx-a11y/anchor-is-valid': 'off',
+    'react/require-default-props': 'off',
   },
   parserOptions: {
     ecmaVersion: 2020,
