+++ src/components/WindowControls.tsx
